Simple creation time propagation for single time step
Need to extend this to an arbitrary skeleton now. In every time step, we
will thus be able to count the number of new segments as well as the
number of old segments.

Ideally, new segments only occur "away" from the origin of the data,
i.e., in direction of the main growth.