Detecting one-to-many matches and many-to-one matches properly
I am now classifying individual *pixels* instead of individual *edges*
in the matching. As expected, the number of growth and decay pixels is
rather small.

Hence, we will probably have to resort to finding connected components
in the bipartite matching graph for all the irregular pixels. I wonder
whether this can be solved by a simple UF structure again...